refactor: rename data.firebase package to data.remote and organize remote models
