+++ composeApp/src/commonMain/kotlin/dev/saragones3/genogramia/data/remote/model/AuthUser.kt
@@ -1,4 +1,4 @@
-package dev.saragones3.genogramia.data.firebase
+package dev.saragones3.genogramia.data.remote.model
 
 data class AuthUser(
     val uid: String,
