+++ composeApp/src/iosMain/kotlin/dev/saragones3/genogramia/KoinHelper.kt
@@ -1,7 +1,7 @@
 package dev.saragones3.genogramia
 
-import dev.saragones3.genogramia.data.firebase.FirebaseAuthDelegate
-import dev.saragones3.genogramia.data.firebase.FirestoreDelegate
+import dev.saragones3.genogramia.data.remote.FirebaseAuthDelegate
+import dev.saragones3.genogramia.data.remote.FirestoreDelegate
 import dev.saragones3.genogramia.di.getSharedModules
 import dev.saragones3.genogramia.di.iosFirebaseModule
 import org.koin.core.context.startKoin
