+++ composeApp/src/iosMain/kotlin/dev/saragones3/genogramia/data/remote/FirebaseProviderImpl.kt
@@ -1,6 +1,6 @@
-package dev.saragones3.genogramia.data.firebase
+package dev.saragones3.genogramia.data.remote
 
-import dev.saragones3.genogramia.data.error.toAuthError
+import dev.saragones3.genogramia.data.remote.model.AuthUser
 import kotlinx.coroutines.suspendCancellableCoroutine
 import kotlin.coroutines.resumeWithException
