+++ composeApp/src/iosMain/kotlin/dev/saragones3/genogramia/di/PlatformModule.kt
@@ -1,11 +1,11 @@
 package dev.saragones3.genogramia.di
 
-import dev.saragones3.genogramia.data.firebase.FirebaseAuthDelegate
-import dev.saragones3.genogramia.data.firebase.FirebaseProvider
-import dev.saragones3.genogramia.data.firebase.FirebaseProviderImpl
-import dev.saragones3.genogramia.data.firebase.FirestoreDelegate
-import dev.saragones3.genogramia.data.firebase.FirestoreProvider
-import dev.saragones3.genogramia.data.firebase.FirestoreProviderImpl
+import dev.saragones3.genogramia.data.remote.FirebaseAuthDelegate
+import dev.saragones3.genogramia.data.remote.FirebaseProvider
+import dev.saragones3.genogramia.data.remote.FirebaseProviderImpl
+import dev.saragones3.genogramia.data.remote.FirestoreDelegate
+import dev.saragones3.genogramia.data.remote.FirestoreProvider
+import dev.saragones3.genogramia.data.remote.FirestoreProviderImpl
 import org.koin.core.module.Module
 import org.koin.dsl.module
