+++ composeApp/src/jvmMain/kotlin/dev/saragones3/genogramia/di/PlatformModule.kt
@@ -1,9 +1,9 @@
 package dev.saragones3.genogramia.di
 
-import dev.saragones3.genogramia.data.firebase.FirebaseProvider
-import dev.saragones3.genogramia.data.firebase.FirebaseProviderImpl
-import dev.saragones3.genogramia.data.firebase.FirestoreProvider
-import dev.saragones3.genogramia.data.firebase.FirestoreProviderImpl
+import dev.saragones3.genogramia.data.remote.FirebaseProvider
+import dev.saragones3.genogramia.data.remote.FirebaseProviderImpl
+import dev.saragones3.genogramia.data.remote.FirestoreProvider
+import dev.saragones3.genogramia.data.remote.FirestoreProviderImpl
 import org.koin.core.module.Module
 import org.koin.dsl.module
