+++ composeApp/src/webMain/kotlin/dev/saragones3/genogramia/data/remote/FirestoreProviderImpl.kt
@@ -1,8 +1,8 @@
 @file:OptIn(kotlin.js.ExperimentalWasmJsInterop::class)
 
-package dev.saragones3.genogramia.data.firebase
+package dev.saragones3.genogramia.data.remote
 
-import dev.saragones3.genogramia.data.firebase.dto.GenogramTreeDto
+import dev.saragones3.genogramia.data.remote.model.GenogramTreeDto
 import kotlinx.serialization.json.Json
 import kotlin.js.JsAny
 import kotlin.js.Promise
